admin expanded functionality
